display the reviews from API
